ADD: data 활용 Assignment 2 : Mock data를 활용하여 여러개의 피트 구현
