+++ src/pages/yoojeong/Main/MainYoojeong.js
@@ -1,32 +1,16 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { ASIDE_BOTTOM } from './AsideBottom';
-import CommentTotalWrap from './CommentTotalWrap';
+import LeftSectionArticle from './LeftSectionArticle';
 import './MainYoojeong.scss';
 
 const MainYoojeong = () => {
-  // input에 작성된 값을 받겠다(배열에 담기 위해)
-  const [inputValue, setInputValue] = useState('');
+  const [feedList, setFeedList] = useState();
 
-  // 작성된 댓글들을 배열에 담아 관리 (array.map 사용을 위해)
-  const [comments, setComments] = useState([]);
-
-  // enter키 누르면 댓글이 comment 배열에 추가
-  const addCommentByEnter = e => {
-    if (e.code === 'Enter' && inputValue.length > 0) {
-      setComments(commentList => [...commentList, inputValue]);
-
-      setInputValue('');
-    }
-  };
-
-  // 게시 버튼 누르면 댓글이 comment 배열에 추가
-  const addCommentByClick = e => {
-    if (inputValue.length > 0) {
-      setComments(commentList => [...commentList, inputValue]);
-
-      setInputValue('');
-    }
-  };
+  useEffect(() => {
+    fetch('/data/yoojeong/feedData.json', { method: 'GET' })
+      .then(res => res.json())
+      .then(data => setFeedList(data));
+  }, []);
 
   return (
     <div className="mainYoojeong">
@@ -83,81 +67,19 @@ const MainYoojeong = () => {
       </nav>
       <div className="mainWrap">
         <div className="leftSection">
-          <div className="leftSectionArticle">
-            <nav className="profileNav">
-              <span className="profileNavInner1">
-                <img
-                  className="profileImg"
-                  src="/images/yoojeong/user.png"
-                  alt="프로필"
+          {feedList &&
+            feedList.map(feed => {
+              return (
+                <LeftSectionArticle
+                  key={feed.id}
+                  profileImg={feed.profileImg}
+                  userId={feed.userId}
+                  articleImg={feed.articleImg}
+                  postText={feed.postText}
+                  postTime={feed.postTime}
                 />
-                <a href="#none" className="profileName">
-                  canon_mj
-                </a>
-              </span>
-              <span className="profileNavInner2">...</span>
-            </nav>
-            <img
-              className="uploadImg"
-              src="/images/yoojeong/chad-montano-MqT0asuoIcU-unsplash.jpg"
-              alt="피자"
-            />
-            <div className="profileIconWrap">
-              <button className="profileHeart">
-                <img
-                  className="profileHeart"
-                  src="/images/yoojeong/heart.png"
-                  alt="하트"
-                />
-              </button>
-              <button className="profileComment">
-                <img
-                  className="profileComment"
-                  src="/images/yoojeong/speech-bubble.png"
-                  alt="댓글"
-                />
-              </button>
-              <button className="profileUproad">
-                <img
-                  className="profileUproad"
-                  src="/images/yoojeong/upload.png"
-                  alt="업로드"
-                />
-              </button>
-              <button className="profileSave">
-                <img
-                  className="profileSave"
-                  src="/images/yoojeong/save-instagram.png"
-                  alt="저장"
-                />
-              </button>
-            </div>
-            <div className="likeWrap">
-              <a href="#none" className="likeId">
-                aineworld
-              </a>
-              님<span className="likeCount">외 10명</span>이 좋아합니다.
-            </div>
-            <ul className="newComment">
-              {comments.map((comment, index) => {
-                return <CommentTotalWrap key={index} comment={comment} />;
-              })}
-            </ul>
-            <div className="commentTime">42분 전</div>
-            <div className="addcommentWrap">
-              <input
-                className="addComment"
-                type="text"
-                placeholder="댓글 달기..."
-                value={inputValue}
-                onChange={e => setInputValue(e.target.value)}
-                onKeyPress={addCommentByEnter}
-              />
-              <button className="addCommentButton" onClick={addCommentByClick}>
-                게시
-              </button>
-            </div>
-          </div>
+              );
+            })}
         </div>
         <div className="rightSection">
           <div className="userProfileWrap">
