+++ src/pages/yoojeong/Main/MainYoojeong.scss
@@ -99,6 +99,7 @@ button {
 
     .leftSectionArticle {
       width: 100%;
+      margin-bottom: 20px;
       background-color: white;
       border: 1.5px solid $gray-border;
 
@@ -236,11 +237,17 @@ button {
           }
         }
       }
-
+      .commentText {
+        margin-left: 14px;
+        margin-bottom: 8px;
+        margin-bottom: 15px;
+      }
       .commentTime {
+        margin-top: 20px;
         margin-left: 14px;
         margin-bottom: 8px;
         color: $gray-text;
+        font-size: 13px;
       }
 
       .addcommentWrap {
@@ -416,6 +423,7 @@ button {
       }
     }
     .etc {
+      margin-top: 40px;
       color: $gray-text;
     }
   }
